Module cincinnati::plugins

The plugin defines and implements the plugin interface, conversion boilerplate, internal plugins, and web plugin helpers

Modules

external
This module references the available external plugins

interface

internal
This module implements the internal plugins

macros
Macro definitions to be used by all plugins

Structs

ExternalIO
Struct used by the InternalPlugin trait impl's

ExternalPluginWrapper
Wrapper struct for a universal implementation of Plugin for all ExternalPlugin implementors

InternalIO
Struct used by the ExternalPlugin trait impl's

InternalPluginWrapper
Wrapper struct for a universal implementation of Plugin for all InternalPlugin implementors

Enums

ExternalError
Error type which corresponds to interface::PluginError

PluginIO
Enum for the two IO variants used by InternalPlugin and ExternalPlugin respectively

PluginResult
Enum for wrapping the interface plugin output types

Traits

ExternalPlugin
Trait to be implemented by external plugins with its native IO type

InternalPlugin
Trait to be implemented by internal plugins with their native IO type

Plugin
Trait which fronts InternalPlugin and ExternalPlugin, allowing their trait objects to live in the same collection

PluginSettings
Settings for a plugin.

Functions

deserialize_config
Validate configuration for a plugin and fill in defaults.

process
Processes all given Plugins sequentially.