Struct cincinnati::Graph

pub struct Graph { /* fields omitted */ }

Graph type which stores Release as node-weights and Empty as edge-weights.

Methods

impl Graph

pub fn add_release<R>(&mut self, release: R) -> Result<ReleaseId, Error> where
R: Into<Release>,

Add a release to the graph.

Fails if the version already exists.

pub fn add_edge(
&mut self,
from: &ReleaseId,
to: &ReleaseId
) -> Result<(), Error>

Add a transition (edge) from source to target.

Fails with the WoulcCycle error if the new edge would lead to a cycle.

pub fn add_edges(
&mut self,
indices: HashMap<ReleaseId, ReleaseId>
) -> Result<(), Error>

Add edges for all given key/value pairs of releases.

pub fn find_by_version(&self, version: &str) -> Option<ReleaseId>

Returns a Some(ReleaseId) if the version exists in the graph, None otherwise.

pub fn remove_edge(
&mut self,
from: &ReleaseId,
to: &ReleaseId
) -> Result<(), Error>

Removes the directed edge between the given releases.

pub fn remove_edges(
&mut self,
indices: HashMap<ReleaseId, ReleaseId>
) -> Result<(), Error>

Remove the directed edges given by the key/value pairs of releases.

pub fn remove_edge_by_index(&mut self, index: EdgeIndex) -> Result<(), Error>

Remove the edge with the given index.

Fails if the edge wasn't found and thus couldn't be removed.

pub fn remove_edges_by_index(
&mut self,
indices: &[EdgeIndex]
) -> Result<(), Error>

Remove the edges with the given indices.

Stops and fails at the first edge which couldn't be removed.

pub fn find_by_fn_mut<F>(&mut self, filter_fn: F) -> Vec<(ReleaseId, String)> where
F: FnMut(&mut Release) -> bool,

Returns tuples of ReleaseId and its version String for releases for which filter_fn returns true.

filter_fn is able to mutate the release as it receives a mutable borrow.

pub fn find_by_metadata_pair(
&self,
key: &str,
value: &str
) -> Vec<(ReleaseId, String)>

Returns tuples of ReleaseId and its version String for releases which match the given metadata key/value pair.

pub fn find_by_metadata_key(
&self,
key: &str
) -> Vec<(ReleaseId, String, String)>

Returns tuples of ReleaseId, its version String, and the value for the given key for releases which match the given metadata key.

pub fn get_metadata_as_ref_mut(
&mut self,
release_id: &ReleaseId
) -> Result<&mut HashMap<String, String>, Error>

Returns a mutable reference to the metadata for the given release.

Important traits for NextReleases<'a>
impl<'a> Iterator for NextReleases<'a> type Item = (EdgeIndex, NodeIndex, &'a Release);

pub fn next_releases(&self, source: &ReleaseId) -> NextReleases

Returns NextReleases for the given release.

NextReleases can be used to iterate over all direct children of the given release.

Important traits for PreviousReleases<'a>
impl<'a> Iterator for PreviousReleases<'a> type Item = (EdgeIndex, NodeIndex, &'a Release);

pub fn previous_releases(&self, source: &ReleaseId) -> PreviousReleases

Returns PreviousReleases for the given release.

PreviousReleases can be used to iterate over all direct parents of the given release.

pub fn releases_count(&self) -> u64

Return the number of releases (nodes) in the graph.

pub fn remove_releases(&mut self, to_remove: Vec<ReleaseId>) -> usize

Removes the nodes with the given ReleaseIds and returns the number of removed releases.

The ReleaseIds are sorted and removed in reverse order. This is required because daggy::Dag::remove_node() shifts the NodeIndices and therefore invalidates all the ones which are higher than the removed one.

pub fn remove_nodes(&mut self, to_remove: Vec<NodeIndex>) -> usize

Removes the nodes with the given NodeIndex and returns the number of removed nodes.

pub fn prune_abstract(&mut self) -> usize

Prune the graph from all abstract releases

Return the number of pruned releases